requirements.txt
Flask==2.3.2
celery==5.3.6
redis==5.0.3
python-dotenv==1.0.1
twilio==8.10.0
